backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
python-dotenv
aiofiles
pytesseract
Pillow
PyPDF2
python-docx
pyhwp
openai-whisper
